+++ Backtrace.Mobile/MainActivity.cs
@@ -14,11 +14,10 @@ namespace Backtrace.Mobile
     {
         protected override void OnCreate(Bundle savedInstanceState)
         {
-            base.OnCreate(savedInstanceState);
             //setup database 
             // we get external storage directory and special directory created for databse - Backtrace directory
             string directoryPath = Path.Combine(Android.OS.Environment.ExternalStorageDirectory.AbsolutePath, "Backtrace");
-            var database = new BacktraceDatabase<object>(new BacktraceDatabaseSettings(directoryPath));
+            var database = new BacktraceDatabase(new BacktraceDatabaseSettings(directoryPath));
 
             //setup client configuration
             var credentials = new BacktraceCredentials(ApplicationSettings.Host, ApplicationSettings.Token);
@@ -26,8 +25,11 @@ namespace Backtrace.Mobile
 
             // Initialize new BacktraceClient 
             BacktraceClient client = new BacktraceClient(clientConfiguration, database);
+
             // Send async report to a server with custom client message
             var result = client.SendAsync("Hello from Xamarin").Result;
+
+            base.OnCreate(savedInstanceState);
             // Set our view from the "main" layout resource
             SetContentView(Resource.Layout.Main);
         }