+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 ﻿# Backtrace C# Release Notes
 
+
+## Version 1.3.2 - 03.07.2018
+- Fixed a invalid type for process.age attribute 
+
+## Version 1.3.1 - 28.06.2018
+- Fixed a double-dispose bug in BacktraceApi
+
+## Version 1.3 - 26.06.2018
+- Attributes dictionary now use an `object` type instead of a generic type for better flexibility,
+- `OnClientReportLimitReached` use `BacktraceReportBase` instead of `BacktraceReport`,
+- `BacktraceResult` store `BacktraceReportBase` instead of `BacktraceReport`. 
+
 ## Version 1.2.3 - 21.06.2018
-- Enum is now available as a primitive value in BacktraceAttributes
+- Enum is now available as a primitive value in BacktraceAttributes.
 
 ## Version 1.2.2 - 19.06.2018
 - Ignore exception object in BacktraceReport in serialization. Change exception conditions in BacktraceReport.